Extend accelerator scheduler with temperature support
