+++ src/accelerator_scheduler.py
@@ -5,9 +5,12 @@ from __future__ import annotations
 import threading
 import time
 from collections import deque
-from typing import Callable, Deque, Dict
+from typing import Callable, Deque, Dict, Optional, TYPE_CHECKING
 
 import psutil
+
+if TYPE_CHECKING:  # pragma: no cover - avoid heavy import at runtime
+    from .telemetry import TelemetryLogger
 try:  # pragma: no cover - optional torch dependency
     import torch  # type: ignore
 except Exception:  # pragma: no cover - allow running without torch
@@ -24,9 +27,25 @@ except Exception:  # pragma: no cover - allow running without torch_xla
 class AcceleratorScheduler:
     """Queue jobs and execute when device utilization is low."""
 
-    def __init__(self, max_util: float = 0.9, check_interval: float = 1.0) -> None:
+    def __init__(
+        self,
+        max_util: float = 0.9,
+        check_interval: float = 1.0,
+        *,
+        max_temp: Optional[float] = None,
+        telemetry: Optional["TelemetryLogger"] = None,
+    ) -> None:
         self.max_util = max_util
         self.check_interval = check_interval
+        self.max_temp = max_temp
+        if telemetry is not None:
+            self.telemetry = telemetry
+        elif max_temp is not None:
+            from .telemetry import TelemetryLogger as _TelemetryLogger
+
+            self.telemetry = _TelemetryLogger(interval=check_interval)
+        else:
+            self.telemetry = None
         self.queues: Dict[str, Deque[Callable[[], None]]] = {
             "cpu": deque(),
             "gpu": deque(),
@@ -43,12 +62,23 @@ class AcceleratorScheduler:
     def _utilization(self, accelerator: str) -> float:
         if accelerator == "gpu" and torch is not None and torch.cuda.is_available():
             try:
-                return (
+                util = (
                     torch.cuda.memory_allocated()
                     / torch.cuda.get_device_properties(0).total_memory
                 )
             except Exception:
-                return 0.0
+                util = 0.0
+            if (
+                self.max_temp is not None
+                and self.telemetry is not None
+                and util < 1.0
+            ):
+                try:
+                    if self.telemetry.gpu_temperature() >= self.max_temp:
+                        return 1.0
+                except Exception:
+                    pass
+            return util
         if accelerator == "tpu" and _HAS_XLA:
             try:
                 info = xm.get_memory_info("xla:0")
