+++ tests/test_thermal_gpu_scheduler.py
@@ -39,8 +39,8 @@ _load('asi.carbon_tracker', 'src/carbon_tracker.py')
 _load('asi.memory_event_detector', 'src/memory_event_detector.py')
 TelemetryLogger = _load('asi.telemetry', 'src/telemetry.py').TelemetryLogger
 accel_mod = _load('asi.accelerator_scheduler', 'src/accelerator_scheduler.py')
-ThermalScheduler = _load('asi.thermal_gpu_scheduler', 'src/thermal_gpu_scheduler.py').ThermalGPUAwareScheduler
 hpc_mod = _load('asi.hpc_schedulers', 'src/hpc_schedulers.py')
+AcceleratorScheduler = accel_mod.AcceleratorScheduler
 
 
 class TestThermalScheduler(unittest.TestCase):
@@ -64,7 +64,7 @@ class TestThermalScheduler(unittest.TestCase):
         temps = [90.0]
         logger = TelemetryLogger(interval=0.05)
         logger.gpu_temperature = lambda index=0: temps[0]
-        sched = ThermalScheduler(max_temp=80.0, check_interval=0.05, telemetry=logger)
+        sched = AcceleratorScheduler(max_temp=80.0, check_interval=0.05, telemetry=logger)
         ran = []
         sched.add(lambda: ran.append(1))
         time.sleep(0.1)
